[#47] Check for active editor, improve reporting
Under some conditions, Atom enables the 'atom-text-editor' 
context-menu, even though the active Pane item is not a 
TextEditor instance. 

In this circumstance, @openEditor can be called, but 
workspace.getActiveTextEditor() will return 'undefined'.

This commit introduces a sanity check against this case
and also changes the flow of control to prevent the error 
from propagating.